chore(*): Refactor to reuse code between selector and pivot. Expose dimensionsSortingExpressionsChange event.
