+++ projects/igniteui-angular/src/lib/grids/common/grid.interface.ts
@@ -29,7 +29,7 @@ import { GridColumnDataType } from '../../data-operations/data-util';
 import { IgxFilteringOperand } from '../../data-operations/filtering-condition';
 import { IColumnPipeArgs, MRLResizeColumnInfo } from '../columns/interfaces';
 import { IgxSummaryResult } from '../summaries/grid-summary';
-import { ISortingExpression, ISortingStrategy } from '../../data-operations/sorting-strategy';
+import { ISortingExpression, ISortingStrategy, SortingDirection } from '../../data-operations/sorting-strategy';
 import { IGridGroupingStrategy, IGridSortingStrategy } from './strategy';
 import { IForOfState, IgxGridForOfDirective } from '../../directives/for-of/for_of.directive';
 import { OverlaySettings } from '../../services/overlay/utilities';
@@ -641,6 +641,7 @@ export interface PivotGridType extends GridType {
     moveDimension(dimension: IPivotDimension, targetCollectionType: PivotDimensionType, index? : number);
     getDimensionsByType(dimension: PivotDimensionType);
     toggleDimension(dimension: IPivotDimension);
+    sortDimension(dimension: IPivotDimension, sortDirection: SortingDirection);
     toggleValue(value: IPivotValue);
     moveValue(value: IPivotValue, index?: number);
     rowDimensionWidthToPixels(dim: IPivotDimension): number;
