+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.component.html
@@ -32,7 +32,7 @@
             class="igx-grid__scroll-on-drag-pinned" [style.left.px]="pinnedWidth"></span>
         <ng-template igxGridFor let-rowData [igxGridForOf]="data
         | pivotGridFilter:pivotConfiguration:filterStrategy:advancedFilteringExpressionsTree:filteringPipeTrigger:pipeTrigger
-        | pivotGridSort:pivotConfiguration:sortStrategy:id:pipeTrigger
+        | pivotGridSort:dimensionsSortingExpressions:sortStrategy:id:pipeTrigger
         | pivotGridRow:pivotConfiguration:expansionStates:pipeTrigger:sortingExpressions
         | pivotGridColumn:pivotConfiguration:expansionStates:pipeTrigger:sortingExpressions
         | pivotGridAutoTransform:pivotConfiguration:pipeTrigger
