+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.component.ts
@@ -63,6 +63,7 @@ import { GridBaseAPIService } from '../api.service';
 import { IgxGridForOfDirective } from '../../directives/for-of/for_of.directive';
 import { IgxPivotRowDimensionContentComponent } from './pivot-row-dimension-content.component';
 import { IgxPivotGridColumnResizerComponent } from '../resizing/pivot-grid/pivot-resizer.component';
+import { ISortingExpression, SortingDirection } from '../../data-operations/sorting-strategy';
 
 let NEXT_ID = 0;
 const MINIMUM_COLUMN_WIDTH = 200;
@@ -101,6 +102,19 @@ export class IgxPivotGridComponent extends IgxGridBaseDirective implements OnIni
      */
     @Output()
     public dimensionsChange = new EventEmitter<IDimensionsChange>();
+
+
+    /**
+     * Emitted when a dimension is sorted.
+     *
+     * @example
+     * ```html
+     * <igx-pivot-grid #grid [data]="localData" [height]="'305px'"
+     *              (dimensionsSortingExpressionsChange)="dimensionsSortingExpressionsChange($event)"></igx-pivot-grid>
+     * ```
+     */
+    @Output()
+    public dimensionsSortingExpressionsChange = new EventEmitter<ISortingExpression[]>();
 
     /**
      * Emitted when the values collection is changed via the grid chip area.
@@ -1352,6 +1366,61 @@ export class IgxPivotGridComponent extends IgxGridBaseDirective implements OnIni
         this.pipeTrigger++;
         this.valuesChange.emit({ values: this.pivotConfiguration.values });
         this.reflow();
+    }
+
+    /**
+     * Sort the dimension and its children in the provided direction.
+     * @example
+     * ```typescript
+     * this.grid.sortDimension(dimension, SortingDirection.Asc);
+     * ```
+     * @param value The value to be toggled.
+     */
+    public sortDimension(dimension: IPivotDimension, sortDirection: SortingDirection) {
+        const dimensionType = this.getDimensionType(dimension);
+        dimension.sortDirection = sortDirection;
+        // apply same sort direction to children.
+        let dim = dimension;
+        while (dim.childLevel) {
+            dim.childLevel.sortDirection = dimension.sortDirection;
+            dim = dim.childLevel;
+        }
+        this.dimensionsSortingExpressions = this.generateDimensionSortingExpressions(this.rowDimensions);
+        this.sortingExpressionsChange.emit(this.dimensionsSortingExpressions);
+        if (dimensionType === PivotDimensionType.Column) {
+            this.setupColumns();
+        }
+        this.cdr.detectChanges();
+    }
+
+    /**
+     * @hidden
+     * @internal
+     */
+    public dimensionsSortingExpressions: Array<ISortingExpression> = [];
+
+    /**
+     * @hidden
+     * @internal
+     */
+    protected generateDimensionSortingExpressions(dimensions: IPivotDimension[]): ISortingExpression[] {
+        const expressions: ISortingExpression[] = [];
+        PivotUtil.flatten(dimensions).forEach(x => {
+            if (x.sortDirection) {
+                expressions.push({
+                    dir: x.sortDirection,
+                    fieldName: x.memberName,
+                    strategy: DefaultPivotSortingStrategy.instance()
+                });
+            } else {
+                expressions.push({
+                    dir: SortingDirection.None,
+                    fieldName: x.memberName,
+                    strategy: DefaultPivotSortingStrategy.instance()
+                });
+            }
+        });
+        return expressions;
     }
 
     public getDimensionsByType(dimension: PivotDimensionType) {
